Configure Watson API request
