+++ server/server.js
@@ -8,12 +8,12 @@ const dev = require('./dev.js');
 const requestHandler = require('./requestHandler.js');
 const callingHandler = require('./calling/callingHandler.js');
 const authHandler = require('./auth/authHandler.js');
-const sentimentHandler = require('./sentiment/sentimentHandler.js');
+const nlpHandler = require('./nlp/nlpHandler.js');
 const Auth = require('./auth/utils.js');
 
 const speechConvertWorker = require('./processing/speechConvertWorker.js');
 const downloadWorker = require('./processing/downloadWorker.js');
-const sentimentWorker = require('./sentiment/sentimentWorker.js');
+const nlpWorker = require('./nlp/nlpWorker.js');
 
 const app = express();
 
@@ -28,7 +28,7 @@ app.use(express.static(path.resolve(__dirname, '../client/public')));
 app.use('/calls', express.static(path.join(__dirname, '/calling/files')));
 
 app.use('/api/calling', callingHandler);
-app.use('/api/sentiment', sentimentHandler);
+app.use('/api/nlp', nlpHandler);
 app.use('/api/auth', authHandler);
 
 app.get('/entries', Auth.authMiddleware, requestHandler.getEntries);
@@ -48,6 +48,6 @@ if (process.env.NODE_ENV !== 'test') {
 speechConvertWorker.start();
 downloadWorker.getFileDetails.start();
 downloadWorker.downloadFiles.start();
-sentimentWorker.start();
+nlpWorker.start();
 
 module.exports = { app };
